requirements.txt
streamlit>=1.28.0
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-auth>=2.23.0
plotly>=5.17.0
pandas>=1.5.0
Pillow>=10.0.0
